refactor(analysis): Split tables.py into modular package
Split large 1869-line tables.py into smaller, more maintainable modules
organized by table purpose:

- summary.py (276 lines): Tables 01, 05 (summary statistics)
- comparison.py (1053 lines): Tables 02, 02b, 04, 06 (statistical comparisons)
- detail.py (602 lines): Tables 03, 07-11 (detailed/appendix tables)
- __init__.py (50 lines): Package exports all functions

Benefits:
- Improved code organization and discoverability
- Easier to navigate and maintain
- Clear separation of concerns by table type
- Reduced cognitive load when working on specific table types

Testing: All 30 existing tests pass without modification.
